add hf state test to structure
